Change long to uint

+++ MachineEmulator/Constants/SizeConstants.cs
@@ -3,6 +3,6 @@
     class SizeConstants
     {
         public const uint MB64 = 2 << 25;
-        public const long RAM_SIZE = 2 << 29;
+        public const uint RAM_SIZE = 2 << 29;
     }
 }
